+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/internal/SyncConfigurationImpl.kt
@@ -17,6 +17,7 @@
 package io.realm.mongodb.internal
 
 import io.realm.internal.InternalConfiguration
+import io.realm.internal.RealmImpl
 import io.realm.internal.interop.RealmConfigurationPointer
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.interop.RealmSyncSessionPointer
@@ -24,20 +25,49 @@ import io.realm.internal.interop.SyncErrorCallback
 import io.realm.internal.interop.sync.PartitionValue
 import io.realm.internal.interop.sync.SyncError
 import io.realm.internal.platform.freeze
+import io.realm.mongodb.exceptions.DownloadingRealmTimeOutException
+import io.realm.mongodb.subscriptions
 import io.realm.mongodb.sync.InitialSubscriptionsCallback
 import io.realm.mongodb.sync.SyncConfiguration
 import io.realm.mongodb.sync.SyncMode
 import io.realm.mongodb.sync.SyncSession
+import io.realm.mongodb.syncSession
+import kotlin.time.Duration
 
+@Suppress("LongParameterList")
 internal class SyncConfigurationImpl(
     private val configuration: io.realm.internal.InternalConfiguration,
     internal val partitionValue: PartitionValue?,
     override val user: UserImpl,
     override val errorHandler: SyncSession.ErrorHandler,
-    val initialSubscriptionsCallback: InitialSubscriptionsCallback?,
-    val rerunInitialSubscriptions: Boolean
+    override val initialSubscriptionsCallback: InitialSubscriptionsCallback?,
+    override val rerunInitialSubscriptions: Boolean,
+    override val shouldWaitForInitialRemoteData: Boolean,
+    override val initialRemoteDataTimeout: Duration
+    ) : InternalConfiguration by configuration, SyncConfiguration {
 
-) : InternalConfiguration by configuration, SyncConfiguration {
+    override suspend fun realmOpened(realm: RealmImpl, fileCreated: Boolean) {
+        initialSubscriptionsCallback?.let { initialSubscriptions ->
+            if (rerunInitialSubscriptions || fileCreated) {
+                realm.subscriptions.update {
+                    with(initialSubscriptions) {
+                        write(realm)
+                    }
+                }
+            }
+        }
+        if (shouldWaitForInitialRemoteData && (fileCreated || rerunInitialSubscriptions)) {
+            val success = if (initialSubscriptionsCallback != null) {
+                realm.subscriptions.waitForSynchronization(initialRemoteDataTimeout)
+            } else {
+                realm.syncSession.downloadAllServerChanges(initialRemoteDataTimeout)
+            }
+            if (!success) {
+                throw DownloadingRealmTimeOutException(this)
+            }
+        }
+        configuration.realmOpened(realm, fileCreated)
+    }
 
     override fun createNativeConfiguration(): RealmConfigurationPointer {
         val ptr = configuration.createNativeConfiguration()