+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 ### Enhancements
 * [Sync] Support for Flexible Sync through `Realm.subscriptions`. (Issue [#824](https://github.com/realm/realm-kotlin/pull/824))
 * [Sync] Added support for `ObjectId` ([#652](https://github.com/realm/realm-kotlin/issues/652)). `ObjectId` can be used as a primary key in model definition.
-* [Sync] Support for `SyncConfiguration.Builder.InitialData()`. (Issue [#422](https://github.com/realm/realm-kotlin/issues/422))
+* [Sync] Support for `SyncConfiguration.Builder.initialData()`. (Issue [#422](https://github.com/realm/realm-kotlin/issues/422))
+* [Sync] Support for `SyncConfiguration.Builder.initialSubscriptions()`.
+* [Sync] Support for `SyncConfiguration.Builder.waitForInitialRemoteData()`. 
 * Support for `RealmConfiguration.Builder.initialData()`. (Issue [#579](https://github.com/realm/realm-kotlin/issues/579))
 
 ### Fixed